GraphBuilder and bug fixes

+++ graphutils.h
@@ -13,7 +13,7 @@ struct Edge{
 
 struct Path{
 	bool success;
-	int total_cost;
+	float total_cost;
 	std::vector<int> indices;
 };
